+++ src/codec_scid4.h
@@ -56,7 +56,7 @@ public: // ICodecDatabase interface
 		return res;
 	};
 
-	const byte* getGameData(uint32_t offset, uint32_t length) override {
+	const byte* getGameData(uint64_t offset, uint32_t length) override {
 		if (offset >= gfile_.size())
 			return NULL;
 		if (length > sizeof(gamecache_))
@@ -121,13 +121,13 @@ public: // CodecNative interface
 	 * data (usable to retrieve the data with getGameData()).
 	 * - on failure, a @e std::pair containing an error code and 0.
 	 */
-	std::pair<errorT, uint32_t> dyn_addGameData(const byte* src,
+	std::pair<errorT, uint64_t> dyn_addGameData(const byte* src,
 	                                            size_t length) {
 		ASSERT(src != 0);
 		const char* data = reinterpret_cast<const char*>(src);
 
 		// The Scid4 format uses 32-bits to store games' offset.
-		size_t offset = gfile_.size();
+		uint64_t offset = gfile_.size();
 		static const uint32_t MAXSZ = std::numeric_limits<uint32_t>::max();
 		if (offset >= MAXSZ || length >= MAXSZ - offset)
 			return std::make_pair(ERROR_OffsetLimit, 0);
@@ -147,7 +147,7 @@ public: // CodecNative interface
 		}
 
 		errorT err = gfile_.append(data, length);
-		return std::make_pair(err, uint32_t(offset));
+		return std::make_pair(err, offset);
 	}
 
 	/**
